// src/generated-twilio-api-models/twilio_numbers_v2.schema.d.ts
export interface components {
  schemas: {
// ...
    "numbers.v2.regulatory_compliance.bundle": {
      /** @description The unique string that we created to identify the Bundle resource. */
      sid?: string | null;
      /** @description The SID of the [Account](https://www.twilio.com/docs/iam/api/account) that created the Bundle resource. */
      account_sid?: string | null;
      /** @description The unique string of a regulation that is associated to the Bundle resource. */
      regulation_sid?: string | null;
      /** @description The string that you assigned to describe the resource. */
      friendly_name?: string | null;
      /** @description The verification status of the Bundle resource. */
      status?: components["schemas"]["bundle_enum_status"];
      /**
       * Format: date-time
       * @description The date and time in GMT in [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format when the resource will be valid until.
       */
      valid_until?: string | null;
      /** @description The email address that will receive updates when the Bundle resource changes status. */
      email?: string | null;
      /**
       * Format: uri
       * @description The URL we call to inform your application of status changes.
       */
      status_callback?: string | null;
      /**
       * Format: date-time
       * @description The date and time in GMT when the resource was created specified in [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format.
       */
      date_created?: string | null;
      /**
       * Format: date-time
       * @description The date and time in GMT when the resource was last updated specified in [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format.
       */
      date_updated?: string | null;
      /**
       * Format: uri
       * @description The absolute URL of the Bundle resource.
       */
      url?: string | null;
      /**
       * Format: uri-map
       * @description The URLs of the Assigned Items of the Bundle resource.
       */
      links?: Record<string, unknown> | null;
    };
// ...
    "numbers.v2.regulatory_compliance.end_user": {
      /** @description The unique string created by Twilio to identify the End User resource. */
      sid?: string | null;
      /** @description The SID of the [Account](https://www.twilio.com/docs/iam/api/account) that created the End User resource. */
      account_sid?: string | null;
      /** @description The string that you assigned to describe the resource. */
      friendly_name?: string | null;
      /** @description The type of end user of the Bundle resource - can be `individual` or `business`. */
      type?: components["schemas"]["end_user_enum_type"];
      /** @description The set of parameters that are the attributes of the End Users resource which are listed in the End User Types. */
      attributes?: unknown;
      /**
       * Format: date-time
       * @description The date and time in GMT when the resource was created specified in [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format.
       */
      date_created?: string | null;
      /**
       * Format: date-time
       * @description The date and time in GMT when the resource was last updated specified in [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format.
       */
      date_updated?: string | null;
      /**
       * Format: uri
       * @description The absolute URL of the End User resource.
       */
      url?: string | null;
    };
// ...
    "numbers.v2.regulatory_compliance.supporting_document": {
      /** @description The unique string created by Twilio to identify the Supporting Document resource. */
      sid?: string | null;
      /** @description The SID of the [Account](https://www.twilio.com/docs/iam/api/account) that created the Document resource. */
      account_sid?: string | null;
      /** @description The string that you assigned to describe the resource. */
      friendly_name?: string | null;
      /** @description The image type uploaded in the Supporting Document container. */
      mime_type?: string | null;
      /** @description The verification status of the Supporting Document resource. */
      status?: components["schemas"]["supporting_document_enum_status"];
      /** @description The failure reason of the Supporting Document Resource. */
      failure_reason?: string | null;
      /** @description The type of the Supporting Document. */
      type?: string | null;
      /** @description The set of parameters that are the attributes of the Supporting Documents resource which are listed in the Supporting Document Types. */
      attributes?: unknown;
      /**
       * Format: date-time
       * @description The date and time in GMT when the resource was created specified in [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format.
       */
      date_created?: string | null;
      /**
       * Format: date-time
       * @description The date and time in GMT when the resource was last updated specified in [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format.
       */
      date_updated?: string | null;
      /**
       * Format: uri
       * @description The absolute URL of the Supporting Document resource.
       */
      url?: string | null;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export interface operations {
// ...
  ListItemAssignment: {
    parameters: {
      query?: {
        /** @description How many resources to return in each list page. The default is 50, and the maximum is 1000. */
        PageSize?: number;
        /** @description The page index. This value is simply for client state. */
        Page?: number;
        /** @description The page token. This is provided by the API. */
        PageToken?: string;
      };
      path: {
        /** @description The unique string that we created to identify the Bundle resource. */
        BundleSid: string;
      };
    };
    responses: {
      /** @description OK */
      200: {
        content: {
          "application/json": {
            results?: components["schemas"]["numbers.v2.regulatory_compliance.bundle.item_assignment"][];
            meta?: {
              /** Format: uri */
              first_page_url?: string;
              /** Format: uri */
              next_page_url?: string | null;
              page?: number;
              page_size?: number;
              /** Format: uri */
              previous_page_url?: string | null;
              /** Format: uri */
              url?: string;
              key?: string;
            };
          };
        };
      };
    };
  };
// ...
}